mvc고민전 기능부터 구현위한 branch
