+++ src/App.js
@@ -1,28 +1,17 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
 import { MESSAGE } from "./const/message";
 import { ERROR } from "./const/error";
-import inputNumbers from "./control/input";
-import progressLotto from "./control/progress";
-import printLottos from "./control/print";
 
 class App {
-  #inputNumbers = new inputNumbers();
-
-  #progressLotto = new progressLotto();
-  
-  #printLottos = new printLottos();
 
   constructor() {
     this.winning = [];
     this.bonus = 0;
     this.count = 0;
   }
+
   async play() {
-    this.getLottoPrice();
-    this.#progressLotto.getMyLottos();
-    this.#inputNumbers.getNumbers();
-    this.#inputNumbers.getBonusNumbers();
-    this.#printLottos.printWinningStatics(sameNumberObject)
+
   }
 
   async getLottoPrice() {
@@ -39,7 +28,94 @@ class App {
   }
 
 // 이후 getNumers에서 입력 받기 전에 몇개 구매한지 출력해야 함
+  async getNumbers(){
+    const winningNumbers = await MissionUtils.Console.readLineAsync(MESSAGE.WINNING_NUMBER);
+    this.winning = this.checkWinningNumbers(winningNumbers);
+  }
 
+  checkWinningNumbers(winningNumbers){
+    const checkWinningNumber = winningNumbers.spilt(",").map(Number);
+
+    if(!winningNumbers.includes(',')) {throw new Error(ERROR.NO_COMMA)};
+    if(winningNumbers.length !== 6) { throw new Error(ERROR.SIX)};
+    this.checkNumberError(checkWinningNumber);
+    return checkWinningNumber;
+  } 
+
+  async getBonusNumbers(){
+    const bonusNumbers = await MissionUtils.Console.readLineAsync(MESSAGE.BONUS_NUMBER);
+   this.bonus = this.checkBonusNumbers(bonusNumbers);
+  }
+  
+  checkBonusNumbers(bonusNumbers){
+    const checkBonusNumber = bonusNumbers.spilt(",").map(Number);
+
+    if(bonusNumbers.includes(',')) {throw new Error(ERROR.INPUT_COMMA)};
+    this.checkNumberError(checkBonusNumber);
+    return checkBonusNumber;
+  }
+
+  checkNumberError(numbers){
+    if(!Number.isInteger(numbers)) {throw new Error(ERROR.INTEGER)};
+    if(numbers < 1) {throw new Error(ERROR.ONE)};
+    if(numbers > 45) {throw new Error(ERROR.FORTY_FIVE)};
+  }
+  
+  getMyLottos() {
+    const arrayNumbers = [];
+  
+    for(let i = 0; i < this.count; i++){
+      const lottoNumber = MissionUtils.Random.pickUniqueNumbersInRange(1,45,6)
+      arrayNumbers.push(lottoNumber.sort());
+    }
+    this.printHowMany(this.count,arrayNumbers);
+    this.getWinningLottos(arrayNumbers);
+  }
+
+  getWinningLottos(arrayNumbers) {
+    const sameNumbersObject = {
+      'three' : 0,
+      'four' : 0,
+      'five' : 0,
+      'six' : 0,
+      'bonus' : 0
+    }
+    for(let i = 0; i < this.count; i++){
+      const newArray = [...arrayNumbers,...this.winning]
+      const sameNumbers = arrayNumbers.filter(newArray);
+      this.checkSameNumbers(sameNumbers,sameNumbersObject)
+    }
+  }
+
+  checkSameNumbers(sameNumbers,sameNumbersObject) {
+    switch (sameNumbers.length) {
+      case 3:
+        sameNumbersObject['three']++
+        break;
+      case 4:
+        sameNumbersObject['four']++
+        break;
+      case 5:
+        sameNumbersObject['five']++
+        break;
+      case 6:
+        sameNumbersObject['six']++
+        break;      
+      default:
+        return sameNumbersObject
+    }
+    if(sameNumbers.length === 5 && arrayNumbers.includes(checkBonusNumber)) {
+      sameNumbersObject['bonus']++
+    }
+    return sameNumbersObject
+  }
+
+  printHowMany(count, arrayNumbers) {
+      MissionUtils.Console.print(`${count}개를 구매했습니다.`);
+      arrayNumbers.forEach(element => {
+        MissionUtils.Console.print(element);
+      })
+  }
 }
 
 export default App;
